fix(eidos,krites,organon,nous,daemon): resolve 6 code quality audit findings
- Add `use std::path::Path` behind tls cfg gate in pylon server (closes #1905)
- Remove stale TODOs referencing closed issues #1140, #1137, #1129 (closes #1906)
- Log swallowed I/O error at warn level in HNSW orphan check (closes #1907)
- Remove `new_unchecked` and `From<String>`/`From<&str>` on eidos ID types,
  replace with validated `TryFrom` impls, add `InvalidId` error variant to
  graphe, migrate all callers to `::new()` with proper error handling (closes #1908)
- Split computer_use.rs (986 lines) into 6 sub-modules and recall.rs (814 lines)
  into 4 sub-modules, all under 350 lines (closes #1911)
- Add PRAGMA busy_timeout = 5000 to daemon TaskStateStore (closes #1914)
